+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 v0.8.0 - TBD
 ------------
+### Changed
+ - [Issue #101](https://github.com/uber/rides-android-sdk/issues/101) LoginManager now uses AccessTokenStorage
 
 v0.7.0 - 11/17/2017
 ------------